Checkerboard done
