+++ scenes/t.rt
@@ -9,8 +9,8 @@ sp 0,-.2,3 1 0,0,255:0,255,0
 sp 0,.3,2.7 .5 0,128,128:128,128,0
 sp 0,0,0 22 128,128,128:0,0,0
 cy 0,0,3 .57735,-.57735,.57735 .2 2 128,128,255:128,0,0
-#co 0,0,3 -.57735,.57735,.57735 45 2 128,128,255:0,128,0
-#co 0,0,3 .57735,.57735,-.57735 45 2 128,128,255:0,0,128
+co 0,0,3 -.57735,.57735,.57735 45 2 128,128,255:0,128,0
+co 0,0,3 .57735,.57735,-.57735 45 2 128,128,255:0,0,128
 cy 0,0.14,2.7 1,0,0 .2 2 128,128,255:0,0,128
 cy 0,1,3 1,0,0 .2 2 255,128,128:0,128,0
-#co 0,0,3 1,0,0 45 2 255,128,128:0,0,0
+co 0,0,3 1,0,0 45 2 255,128,128:0,0,0
