+++ src/model/impact_color1.c
@@ -6,7 +6,7 @@
 /*   By: parden <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/14 17:39:57 by parden            #+#    #+#             */
-/*   Updated: 2025/06/16 16:37:17 by parden           ###   ########.fr       */
+/*   Updated: 2025/06/16 17:17:41 by parden           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,8 +37,8 @@ t_rgb	color_disk(t_model *m, t_impact *imp, t_dsk *obj)
 	if (!obj->check)
 		return (obj->col1);
 	vec_sub(vec, imp->pos, obj->pos);
-	f2 = roundf(vec_dot(obj->e2, vec));
-	f3 = roundf(vec_dot(obj->e3, vec));
+	f2 = roundf(2 * vec_dot(obj->e2, vec) / obj->rad);
+	f3 = roundf(2 * vec_dot(obj->e3, vec) / obj->rad);
 	if ((f2 + f3) % 2)
 		return (obj->col1);
 	return (obj->col2);
@@ -61,24 +61,6 @@ t_rgb	color_sphere(t_model *m, t_impact *imp, t_sph *obj)
 	return (obj->col2);
 }
 
-t_rgb	color_cyl(t_model *m, t_impact *imp, t_cyl *obj)
-{
-	t_v3	vec;
-	int		f2;
-	int		f3;
-
-	if (!obj->check)
-		return (obj->col1);
-	vec_sub(vec, imp->pos, obj->pos);
-	f2 = roundf(6.5f * vec_dot(vec, obj->ax) / obj->hgt);
-	vec_normalize(vec);
-	f3 = roundf(4 * atan2f(vec_dot(vec, obj->e2), vec_dot(vec, obj->e3)) / M_PI);
-	if ((f2 + f3) % 2)
-		return (obj->col1);
-	return (obj->col2);
-}
-
-
 void	impact_color(t_model *m, t_impact *imp)
 {
 	t_obj	*obj;
@@ -91,7 +73,7 @@ void	impact_color(t_model *m, t_impact *imp)
 	if (obj->type == CYL)
 		imp->col = color_cyl(m, imp, &obj->cyl);
 	if (obj->type == CON)
-		imp->col = obj->con.col1;
+		imp->col = color_con(m, imp, &obj->con);
 	if (obj->type == DSK)
 		imp->col = color_disk(m, imp, &obj->dsk);
 }
